working import

+++ MusicFinder.CLI/Program.cs
@@ -1,7 +1,6 @@
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Configuration;
-using FluentMigrator.Runner;
 using System.Reflection;
 using MusicFinder.Models.Settings;
 using Microsoft.Data.Sqlite;
@@ -10,43 +9,47 @@ using ActionEnum = MusicFinder.Models.Enums.Action;
 using MusicFinder.Actions;
 using MusicFinder.Actions.CLI;
 using MusicFinder.Models;
+using Microsoft.EntityFrameworkCore;
 
 var builder = Host.CreateDefaultBuilder(args)
     .ConfigureServices((hostContext, services) =>
     {
-        var settings = hostContext.Configuration.GetSection("MusicFinder").Get<MusicFinderSettings>() ?? new MusicFinderSettings();
+        services.Configure<MusicFinderSettings>(hostContext.Configuration.GetSection("MusicFinder"));
 
-        var dataDir = string.IsNullOrEmpty(settings.DataDirectory) ? Xdg.Directories.BaseDirectory.DataHome ?? "." : settings.DataDirectory;
-        var musicDir = Path.Combine(dataDir, Assembly.GetExecutingAssembly().GetName().Name ?? "MusicFinder");
-        Directory.CreateDirectory(musicDir);
-        var dbPath = Path.Combine(musicDir, settings.DatabaseName);
-        File.Create(dbPath).Close();
-        var connectionString = new SqliteConnectionStringBuilder { DataSource = dbPath }.ToString();
+        services.AddDbContext<MusicFinderContext>(
+            options =>
+            {
+                var settings = hostContext.Configuration.GetSection("MusicFinder").Get<MusicFinderSettings>() ?? new MusicFinderSettings();
 
-        services.AddSingleton(Options.Create(settings));
-        services.AddSingleton(settings);
-        services.AddSingleton(sp => connectionString);
-        services.AddScoped(sp => new MusicFinderContext());
+                var dataDir = string.IsNullOrEmpty(settings.DataDirectory) ? Xdg.Directories.BaseDirectory.DataHome ?? "." : settings.DataDirectory;
+                var musicDir = Path.Combine(dataDir, Assembly.GetExecutingAssembly().GetName().Name ?? "MusicFinder");
+                Directory.CreateDirectory(musicDir);
+                var dbPath = Path.Combine(musicDir, settings.DatabaseName);
+                if (!File.Exists(dbPath))
+                {
+                    File.Create(dbPath).Close();
+                }
+                var connectionString = new SqliteConnectionStringBuilder { DataSource = dbPath }.ToString();
+                options.UseSqlite(connectionString);
+            }
+        );
+
         services.AddScoped<Import>();
     });
 
 using var host = builder.Build();
 
-using var scope = host.Services.CreateScope();
+var db = host.Services.GetRequiredService<MusicFinderContext>();
+await db.Database.MigrateAsync();
 
-var config = scope.ServiceProvider.GetRequiredService<IOptions<MusicFinderSettings>>();
-var migrator = scope.ServiceProvider.GetRequiredService<IMigrationRunner>();
-migrator.MigrateUp();
-
-var importAction = scope.ServiceProvider.GetRequiredService<Import>();
-
-Console.WriteLine($"Action: {config.Value.Action}, Path: {config.Value.ImportPath}");
+var config = host.Services.GetRequiredService<IOptions<MusicFinderSettings>>();
 
 if (Enum.TryParse(config.Value.Action, out ActionEnum action))
 {
     switch (action)
     {
         case ActionEnum.Import:
+            var importAction = host.Services.GetRequiredService<Import>();
             await importAction.RunAsync(CancellationToken.None);
             break;
         case ActionEnum.Help:
